Factor up to a base class for combat creators
